requirements.txt
# === Core ===
python-dotenv>=1.0.0

# === Database ===
psycopg2-binary>=2.9

# === HTTP Client ===
httpx>=0.27

# === API Server ===
fastapi>=0.115
uvicorn>=0.32
websockets>=13.0
